consolidate original resource request code
